# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for VVSU Tourism & Creative Industries immersive educational platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quests/{id}:
    get:
      operationId: getQuest
      tags: [quests]
      summary: Get quest detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quest detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quest"
  /quests/{id}/submit:
    post:
      operationId: submitQuest
      tags: [quests]
      summary: Submit quest answer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestSubmission"
      responses:
        "200":
          description: Submission result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestResult"
...
  /community/posts:
    get:
      operationId: listCommunityPosts
      tags: [community]
      summary: List community forum posts
      responses:
        "200":
          description: Community posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommunityPost"
    post:
      operationId: createCommunityPost
      tags: [community]
      summary: Create a community post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommunityPostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPost"
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registered user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
...
  /profile:
    get:
      operationId: getMyProfile
      tags: [profile]
      summary: Get my full profile with stats
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    patch:
      operationId: updateMyProfile
      tags: [profile]
      summary: Update my profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
  /profile/{id}:
    get:
      operationId: getPublicProfile
      tags: [profile]
      summary: Get public profile of a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
...
  /admin/courses:
    post:
      operationId: adminCreateCourse
      tags: [admin]
      summary: Create a new course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseRequest"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /admin/quests:
    post:
      operationId: adminCreateQuest
      tags: [admin]
      summary: Create a new quest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestRequest"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quest"
...
components:
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        role:
          type: string
          description: "Student role: guide, marketer, designer, operator"
        stage:
          type: string
          description: "Journey stage name (e.g. Порт отправления)"
        imageUrl:
          type: string
        totalModules:
          type: integer
        completedModules:
          type: integer
        xpReward:
          type: integer
        isLocked:
          type: boolean
        category:
          type: string
      required: [id, title, description, role, stage, totalModules, completedModules, xpReward, isLocked, category]
...
    Quest:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          description: "route, budget, marketing, design"
        difficulty:
          type: string
          description: "easy, medium, hard"
        xpReward:
          type: integer
        isCompleted:
          type: boolean
        imageUrl:
          type: string
        locationName:
          type: string
      required: [id, title, description, type, difficulty, xpReward, isCompleted, locationName]
    QuestSubmission:
      type: object
      properties:
        answer:
          type: string
      required: [answer]
    QuestResult:
      type: object
      properties:
        success:
          type: boolean
        xpEarned:
          type: integer
        feedback:
          type: string
        newAchievements:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"
      required: [success, xpEarned, feedback, newAchievements]
    Achievement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        category:
          type: string
          description: "exploration, learning, social, mastery"
        isUnlocked:
          type: boolean
        unlockedAt:
          type: string
          nullable: true
        xpValue:
          type: integer
      required: [id, title, description, icon, category, isUnlocked, xpValue]
...
    CommunityPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        authorName:
          type: string
        authorRole:
          type: string
        createdAt:
          type: string
        likes:
          type: integer
        replies:
          type: integer
        category:
          type: string
      required: [id, title, content, authorName, authorRole, createdAt, likes, replies, category]
    CreateCommunityPostBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
      required: [title, content, category]
...
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          description: "student or admin"
        studentRole:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        level:
          type: integer
        xp:
          type: integer
      required: [id, email, name, role, studentRole, level, xp]
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6
        name:
          type: string
          minLength: 2
        studentRole:
          type: string
      required: [email, password, name]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        studentRole:
          type: string
    ProfileActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        xpEarned:
          type: integer
        createdAt:
          type: string
      required: [id, type, description, xpEarned, createdAt]
    QuizAttemptHistory:
      type: object
      properties:
        id:
          type: integer
        quizId:
          type: integer
        quizTitle:
          type: string
        score:
          type: integer
        total:
          type: integer
        xpEarned:
          type: integer
        passed:
          type: boolean
        createdAt:
          type: string
      required: [id, quizId, quizTitle, score, total, xpEarned, passed, createdAt]
    CompletedCourse:
      type: object
      properties:
        courseId:
          type: integer
        title:
          type: string
        completedModules:
          type: integer
        totalModules:
          type: integer
      required: [courseId, title, completedModules, totalModules]
    UnlockedAchievement:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        iconType:
          type: string
        category:
          type: string
        xpReward:
          type: integer
        unlockedAt:
          type: string
      required: [id, name, description, iconType, category, xpReward, unlockedAt]
    Profile:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        nextLevelXp:
          type: integer
        currentLevelXp:
          type: integer
        completedQuests:
          type: integer
        completedQuizzes:
          type: integer
        completedModules:
          type: integer
        unlockedAchievements:
          type: array
          items:
            $ref: "#/components/schemas/UnlockedAchievement"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ProfileActivityItem"
        quizHistory:
          type: array
          items:
            $ref: "#/components/schemas/QuizAttemptHistory"
        completedCourses:
          type: array
          items:
            $ref: "#/components/schemas/CompletedCourse"
      required:
        [user, nextLevelXp, currentLevelXp, completedQuests, completedQuizzes,
         completedModules, unlockedAchievements, recentActivity, quizHistory, completedCourses]
    PublicUser:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        role: { type: string }
        studentRole: { type: string }
        bio: { type: string, nullable: true }
        avatarUrl: { type: string, nullable: true }
        level: { type: integer }
        xp: { type: integer }
      required: [id, name, role, studentRole, level, xp]
    PublicProfile:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/PublicUser"
        completedQuests:
          type: integer
        completedQuizzes:
          type: integer
        completedModules:
          type: integer
        unlockedAchievements:
          type: array
          items:
            $ref: "#/components/schemas/UnlockedAchievement"
      required: [user, completedQuests, completedQuizzes, completedModules, unlockedAchievements]
...
    CreateCourseRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        role:
          type: string
        stage:
          type: string
        category:
          type: string
        xpReward:
          type: integer
        imageUrl:
          type: string
      required: [title, description, role, stage, category]
    CreateQuestRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
        difficulty:
          type: string
        location:
          type: string
        xpReward:
          type: integer
        timeEstimate:
          type: integer
      required: [title, description, type, difficulty, location]
...